requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
babelfish==0.6.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.115.14
guessit==3.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
RapidFuzz==3.13.0
rebulk==3.2.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
flask